TRUNK-5992 - Order Service should always set autoExpireDate if duration is set
